moved validate_forecast to its own file
